08-ai-embedded-flexible-energy-grid/dataflows-code/c2d_pipeline/requirements.txt
pandas
numpy
ai-core-sdk
requests
hana_ml
azure-iot-device
azure-iot-hub
shapely